sdl: Integrate SDL into DDL.
Make it possible to specify `MIGRATION TO { [SDL...] }` as a DDL
command. Stop using the `TO eschema $$...$$` format.
