+++ edb/edgeql/parser/grammar/ddl.py
@@ -19,8 +19,6 @@
 
 import collections
 import re
-import sys
-import types
 import typing
 
 from edb.errors import EdgeQLSyntaxError
@@ -31,12 +29,21 @@ from edb.edgeql import qltypes
 from edb.common import parsing, context
 from edb.common.parsing import ListNonterm
 
-from .expressions import Nonterm
-from . import tokens
+from . import expressions
+from . import commondl
 
 from .precedence import *  # NOQA
 from .tokens import *  # NOQA
-from .expressions import *  # NOQA
+from .commondl import *  # NOQA
+from .sdl import *  # NOQA
+
+
+Nonterm = expressions.Nonterm
+Semicolons = commondl.Semicolons
+
+
+sdl_nontem_helper = commondl.NewNontermHelper(__name__)
+_new_nonterm = sdl_nontem_helper._new_nonterm
 
 
 class DDLStmt(Nonterm):
@@ -196,23 +203,6 @@ class InnerDDLStmt(Nonterm):
         self.val = kids[0].val
 
 
-class Semicolons(Nonterm):
-    # one or more semicolons
-    def reduce_SEMICOLON(self, tok):
-        self.val = tok
-
-    def reduce_Semicolons_SEMICOLON(self, *kids):
-        self.val = kids[0].val
-
-
-class OptSemicolons(Nonterm):
-    def reduce_Semicolons(self, *kids):
-        self.val = kids[0].val
-
-    def reduce_empty(self):
-        self.val = None
-
-
 class InnerDDLStmtBlock(ListNonterm, element=InnerDDLStmt,
                         separator=Semicolons):
     pass
@@ -233,20 +223,6 @@ class UnqualifiedPointerName(Nonterm):
                 'unexpected fully-qualified name',
                 context=kids[0].val.context)
         self.val = kids[0].val
-
-
-def _new_nonterm(clsname, clsdict={}, clskwds={}, clsbases=(Nonterm,)):
-    mod = sys.modules[__name__]
-
-    def clsexec(ns):
-        ns['__module__'] = __name__
-        for k, v in clsdict.items():
-            ns[k] = v
-        return ns
-
-    cls = types.new_class(clsname, clsbases, clskwds, clsexec)
-    setattr(mod, clsname, cls)
-    return cls
 
 
 class ProductionHelper:
@@ -374,19 +350,6 @@ commands_block(
     opt=False)
 
 
-class Extending(Nonterm):
-    def reduce_EXTENDING_SimpleTypeNameList(self, *kids):
-        self.val = kids[1].val
-
-
-class OptExtending(Nonterm):
-    def reduce_Extending(self, *kids):
-        self.val = kids[0].val
-
-    def reduce_empty(self, *kids):
-        self.val = []
-
-
 class AlterAbstract(Nonterm):
     def reduce_DROP_ABSTRACT(self, *kids):
         self.val = qlast.SetSpecialField(name='is_abstract', value=False)
@@ -461,6 +424,38 @@ class OptDeltaTarget(Nonterm):
 #
 
 
+class SDLCommandBlock(Nonterm):
+    # this command block can be empty
+    def reduce_LBRACE_OptSemicolons_RBRACE(self, *kids):
+        self.val = qlast.Schema(declarations=[])
+
+    def reduce_statement_without_semicolons(self, *kids):
+        r"""%reduce LBRACE \
+                OptSemicolons SDLShortStatement \
+            RBRACE
+        """
+        self.val = qlast.Schema(declarations=[kids[2].val])
+
+    def reduce_statements_without_optional_trailing_semicolons(self, *kids):
+        r"""%reduce LBRACE \
+                OptSemicolons SDLStatements \
+                OptSemicolons SDLShortStatement \
+            RBRACE
+        """
+        self.val = qlast.Schema(declarations=kids[2].val + [kids[4].val])
+
+    def reduce_LBRACE_OptSemicolons_SDLStatements_RBRACE(self, *kids):
+        self.val = qlast.Schema(declarations=kids[2].val)
+
+    def reduce_statements_without_optional_trailing_semicolons2(self, *kids):
+        r"""%reduce LBRACE \
+                OptSemicolons SDLStatements \
+                Semicolons \
+            RBRACE
+        """
+        self.val = qlast.Schema(declarations=kids[2].val)
+
+
 class CreateDeltaStmt(Nonterm):
     def _parse_schema_decl(self, tok):
         from edb.common.exceptions import get_context
@@ -478,28 +473,14 @@ class CreateDeltaStmt(Nonterm):
             context.rebase_ast_context(ctx, node)
             return node
 
-    def reduce_CreateDelta_TO(self, *kids):
+    def reduce_CreateDelta_SDL(self, *kids):
         r"""%reduce CREATE MIGRATION NodeName \
-                    OptDeltaParents OptDeltaTarget \
+                    OptDeltaParents TO SDLCommandBlock
         """
-        if kids[4].val is None:
-            lang = target = None
-        else:
-            lang, target = kids[4].val
-
-        # currently we only support one valid language for migration target
-        #
-        if lang.val.lower() == 'eschema':
-            target = self._parse_schema_decl(target)
-        else:
-            raise EdgeQLSyntaxError(f'unknown migration language: {lang.val}',
-                                    context=lang.context)
-
         self.val = qlast.CreateDelta(
             name=kids[2].val,
             parents=kids[3].val,
-            language=lang.val.lower(),
-            target=target,
+            target=kids[5].val,
         )
 
     def reduce_CreateDelta_Commands(self, *kids):
@@ -692,33 +673,12 @@ class DropConstraintStmt(Nonterm):
         )
 
 
-class OnExpr(Nonterm):
-    def reduce_ON_LPAREN_Expr_RPAREN(self, *kids):
-        self.val = kids[2].val
-
-
-class OptOnExpr(Nonterm):
-    def reduce_empty(self, *kids):
-        self.val = None
-
-    def reduce_OnExpr(self, *kids):
-        self.val = kids[0].val
-
-
 class OptDelegated(Nonterm):
     def reduce_DELEGATED(self, *kids):
         self.val = True
 
     def reduce_empty(self):
         self.val = False
-
-
-class OptConcreteConstraintArgList(Nonterm):
-    def reduce_LPAREN_OptPosCallArgList_RPAREN(self, *kids):
-        self.val = kids[1].val
-
-    def reduce_empty(self):
-        self.val = []
 
 
 class CreateConcreteConstraintStmt(Nonterm):
@@ -1492,154 +1452,14 @@ class DropModuleStmt(Nonterm):
 #
 # FUNCTIONS
 #
-
-
-class OptDefault(Nonterm):
-    def reduce_empty(self):
-        self.val = None
-
-    def reduce_EQUALS_Expr(self, *kids):
-        self.val = kids[1].val
-
-
-class OptParameterKind(Nonterm):
-
-    def reduce_empty(self):
-        self.val = qltypes.ParameterKind.POSITIONAL
-
-    def reduce_VARIADIC(self, kid):
-        self.val = qltypes.ParameterKind.VARIADIC
-
-    def reduce_NAMEDONLY(self, *kids):
-        self.val = qltypes.ParameterKind.NAMED_ONLY
-
-
-class FuncDeclArgName(Nonterm):
-
-    def reduce_Identifier(self, dp):
-        self.val = dp.val
-        self.context = dp.context
-
-    def reduce_DOLLAR_AnyIdentifier(self, dk, dp):
-        raise EdgeQLSyntaxError(
-            f"function parameters do not need a $ prefix, "
-            f"rewrite as '{dp.val}'",
-            context=dk.context)
-
-    def reduce_DOLLAR_ICONST(self, dk, di):
-        raise EdgeQLSyntaxError(
-            f'numeric parameters are not supported',
-            context=dk.context)
-
-
-class FuncDeclArg(Nonterm):
-    def reduce_kwarg(self, *kids):
-        r"""%reduce OptParameterKind FuncDeclArgName COLON \
-                OptTypeQualifier FullTypeExpr OptDefault \
-        """
-        self.val = qlast.FuncParam(
-            kind=kids[0].val,
-            name=kids[1].val,
-            typemod=kids[3].val,
-            type=kids[4].val,
-            default=kids[5].val
-        )
-
-    def reduce_OptParameterKind_FuncDeclArgName_OptDefault(self, *kids):
-        raise EdgeQLSyntaxError(
-            f'missing type declaration for the `{kids[1].val}` parameter',
-            context=kids[1].context)
-
-
-class FuncDeclArgList(ListNonterm, element=FuncDeclArg,
-                      separator=tokens.T_COMMA):
-    pass
-
-
-class CreateFunctionArgs(Nonterm):
-    def reduce_LPAREN_RPAREN(self, *kids):
-        self.val = []
-
-    def reduce_LPAREN_FuncDeclArgList_RPAREN(self, *kids):
-        args = kids[1].val
-
-        last_pos_default_arg = None
-        last_named_arg = None
-        variadic_arg = None
-        names = set()
-        for arg in args:
-            if arg.name in names:
-                raise EdgeQLSyntaxError(
-                    f'duplicate parameter name `{arg.name}`',
-                    context=arg.context)
-            names.add(arg.name)
-
-            if arg.kind is qltypes.ParameterKind.VARIADIC:
-                if variadic_arg is not None:
-                    raise EdgeQLSyntaxError(
-                        'more than one variadic argument',
-                        context=arg.context)
-                elif last_named_arg is not None:
-                    raise EdgeQLSyntaxError(
-                        f'NAMED ONLY argument `{last_named_arg.name}` '
-                        f'before VARIADIC argument `{arg.name}`',
-                        context=last_named_arg.context)
-                else:
-                    variadic_arg = arg
-
-                if arg.default is not None:
-                    raise EdgeQLSyntaxError(
-                        f'VARIADIC argument `{arg.name}` '
-                        f'cannot have a default value',
-                        context=arg.context)
-
-            elif arg.kind is qltypes.ParameterKind.NAMED_ONLY:
-                last_named_arg = arg
-
-            else:
-                if last_named_arg is not None:
-                    raise EdgeQLSyntaxError(
-                        f'positional argument `{arg.name}` '
-                        f'follows NAMED ONLY argument `{last_named_arg.name}`',
-                        context=arg.context)
-
-                if variadic_arg is not None:
-                    raise EdgeQLSyntaxError(
-                        f'positional argument `{arg.name}` '
-                        f'follows VARIADIC argument `{variadic_arg.name}`',
-                        context=arg.context)
-
-            if arg.kind is qltypes.ParameterKind.POSITIONAL:
-                if arg.default is None:
-                    if last_pos_default_arg is not None:
-                        raise EdgeQLSyntaxError(
-                            f'positional argument `{arg.name}` without '
-                            f'default follows positional argument '
-                            f'`{last_pos_default_arg.name}` with default',
-                            context=arg.context)
-                else:
-                    last_pos_default_arg = arg
-
-        self.val = args
-
-
-def _parse_language(node):
-    try:
-        return qlast.Language(node.val.upper())
-    except ValueError:
-        raise EdgeQLSyntaxError(
-            f'{node.val} is not a valid language',
-            context=node.context) from None
-
-
 class FromFunction(Nonterm):
     def reduce_FROM_Identifier_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         code = kids[2].val.value
         self.val = qlast.FunctionCode(language=lang, code=code)
 
     def reduce_FROM_Identifier_FUNCTION_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM FUNCTION clause',
@@ -1649,7 +1469,7 @@ class FromFunction(Nonterm):
                                       from_function=kids[3].val.value)
 
     def reduce_FROM_Identifier_EXPRESSION(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM clause',
@@ -1730,17 +1550,6 @@ commands_block(
 )
 
 
-class OptTypeQualifier(Nonterm):
-    def reduce_SET_OF(self, *kids):
-        self.val = qltypes.TypeModifier.SET_OF
-
-    def reduce_OPTIONAL(self, *kids):
-        self.val = qltypes.TypeModifier.OPTIONAL
-
-    def reduce_empty(self):
-        self.val = qltypes.TypeModifier.SINGLETON
-
-
 class CreateFunctionStmt(Nonterm, _ProcessFunctionBlockMixin):
     def reduce_CreateFunction(self, *kids):
         r"""%reduce CREATE FUNCTION NodeName CreateFunctionArgs \
@@ -1762,11 +1571,6 @@ class DropFunctionStmt(Nonterm):
         self.val = qlast.DropFunction(
             name=kids[2].val,
             params=kids[3].val)
-
-
-class FunctionType(Nonterm):
-    def reduce_FullTypeExpr(self, *kids):
-        self.val = kids[0].val
 
 
 #
@@ -1791,7 +1595,7 @@ class OperatorKind(Nonterm):
 class OperatorCode(Nonterm):
 
     def reduce_FROM_Identifier_OPERATOR_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM OPERATOR clause',
@@ -1813,7 +1617,7 @@ class OperatorCode(Nonterm):
             language=lang, from_operator=sql_operator)
 
     def reduce_FROM_Identifier_FUNCTION_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM FUNCTION clause',
@@ -1823,7 +1627,7 @@ class OperatorCode(Nonterm):
                                       from_function=kids[3].val.value)
 
     def reduce_FROM_Identifier_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM clause',
@@ -1833,7 +1637,7 @@ class OperatorCode(Nonterm):
                                       code=kids[2].val.value)
 
     def reduce_FROM_Identifier_EXPRESSION(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM clause',
@@ -1998,7 +1802,7 @@ class CastAllowedUse(Nonterm):
 class CastCode(Nonterm):
 
     def reduce_FROM_Identifier_FUNCTION_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang not in {qlast.Language.SQL, qlast.Language.EdgeQL}:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM FUNCTION clause',
@@ -2008,7 +1812,7 @@ class CastCode(Nonterm):
                                   from_function=kids[3].val.value)
 
     def reduce_FROM_Identifier_BaseStringConstant(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang not in {qlast.Language.SQL, qlast.Language.EdgeQL}:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM clause',
@@ -2018,7 +1822,7 @@ class CastCode(Nonterm):
                                   code=kids[2].val.value)
 
     def reduce_FROM_Identifier_CAST(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM CAST clause',
@@ -2027,7 +1831,7 @@ class CastCode(Nonterm):
         self.val = qlast.CastCode(language=lang, from_cast=True)
 
     def reduce_FROM_Identifier_EXPRESSION(self, *kids):
-        lang = _parse_language(kids[1])
+        lang = commondl._parse_language(kids[1])
         if lang != qlast.Language.SQL:
             raise EdgeQLSyntaxError(
                 f'{lang} language is not supported in FROM EXPRESSION clause',
