+++ edb/testbase/server.py
@@ -362,7 +362,7 @@ class DatabaseTestCase(ClusterTestCase, ConnectedTestCaseMixin):
                     script += f'\nCREATE MODULE {module_name};'
 
                 script += f'\nCREATE MIGRATION {module_name}::d1'
-                script += f' TO eschema $${schema}$$;'
+                script += f' TO {{ {schema} }};'
                 script += f'\nCOMMIT MIGRATION {module_name}::d1;'
 
         if cls.SETUP:
