+++ tests/test_edgeql_datatypes.py
@@ -24,7 +24,7 @@ from edb.testbase import server as tb
 
 class TestEdgeQLDT(tb.QueryTestCase):
     SETUP = '''
-        CREATE MIGRATION default::m TO eschema $$
+        CREATE MIGRATION default::m TO {
             scalar type seq_t extending sequence;
             scalar type seq2_t extending sequence;
 
@@ -35,7 +35,7 @@ class TestEdgeQLDT(tb.QueryTestCase):
             type Obj2 {
                 property seq_prop -> seq2_t;
             };
-        $$;
+        };
 
         COMMIT MIGRATION default::m;
     '''
